DiabetesMapDescription
Why are there HTML files available?


The first file contains a world map showing randomly distributed individuals, along with their ages and whether they have diabetes or not.

The second HTML file is created to enhance the visibility of diabetes patients on the map. It includes features that highlight their locations more clearly, allowing for better analysis and interpretation of the data. This file serves to provide an interactive experience for users, enabling them to explore the geographic distribution of diabetes patients effectively.