Basic level EDA, Visualizing raw data
In this section of my project, I performed basic Exploratory Data Analysis (EDA) and visualized the data to make it more understandable. This serves as a preliminary step for future processes such as model training and performance optimization, helping to interpret the raw data and prepare it for further analysis.