Enhanced EDA Analysis, Feature Engineering
The third part, which I added later to my project, is dedicated to making the EDA analyses more realistic and meaningful for real-life applications. In this section, I fully cleaned the dataset, identified missing and incorrect data, and worked on finding the most efficient scaling techniques. In the final part, I included a section on feature engineering and feature importance, explaining how more meaningful analyses can be performed with the added features.